Add files via upload
Fix in function 'get_subtype_by_recvq'. Previously index error for systems with 1 or 2 queues. Additionally, the plot was improved for systems with an uneven number of queues.
Q Apply latency KPIs "by day" changed from MAX to AVG.
Connect to database moved to a function in the library to clean up the code,
More documentation to explain the displayed performance plots.